Updates `sic` to use build.rs for releases instead of cargo-make.
Previously, cargo-make (`cargo make release`) was used to build release builds.
This worked for separately build binaries; but not for binaries installed
through cargo (`cargo install --force sic`).
Additionally, on Windows, it required some tools like `touch` to be
pre-installed.

To overcome these two issues, this commit includes a build.rs file which will
be run before each build.

The build script checks whether cargo-bom is installed; and if not,
install it using cargo.

Then it will proceed and run cargo-bom on this project, which will be written
to a DEP_LICENSES file in the "target/" folder (preventing it to be included
in the source repo).

Then the actual build will proceed which relies on this file to exist.