Fix bug where a script option was expected for all image inputs.
The pipeline executes all steps in a linear fashion.
One con is that each step has to manually check whether it should actually run.
This check did not happen for ImageOperationsProcessor.
Therefore every run which didn't exit before running ImageOperationsProcessor's process(),
expected --script VALUE to be specified correctly.
This commit fixes that issue by checking whether it was defined first.
Perhaps the "pipeline" should be made more formal in the future, so that
steps will only have to run when it would be valid to do so.