Removed cli runner test cases.
Test cases have been moved to https://github.com/foresterre/sic_cli_runner_tests

The reason is that the tool used for these test cases is a prototype which is not public (yet).
I don't want to pollute this repo with some of my experiments if they don't relate too closely
with sic =).