This commit adds  "color type preprocessing".
This is the case if the output image format does not support the color type held by the image buffer prior
to the final conversion.

The open question remains as follows: does a user expect for an image to be able to convert to a
format even if the color type is not supported?

And even if the user does, should we?

I suspect that users expect that color type conversions should happen automatically.

From this commit onward, until ansers for the questions above are established,
the color type will adapted for PNM bitmap (to Luma8), PNM graymap (to Luma8) and
for PNM pixmap (to Rgb8).

It is however possible to disable this automatic adjustment by providing the
(awfully long) --disable-automatic-color-type-adjustment flag.