Updated pest from version '1.x.x' to '2.x.x'.
The handled breaking changes are:
- 'operation_sequence' now handles whitespace explicitly.
- EOI is no longer silent, so the rule will be filtered out before

Additionally, test_setup from the operations module is renamed to mod_test_includes to be
in line with the other test includes module.

A few lines have been moved within the same module (in particular the filter3x3 parsing code), for
improved readability.

Future:
- Looking at the pest 2.0.0 release highlights, 'const _GRAMMER' might no longer be
necessary to refresh the included pest grammar. It is not too clear from the docs
whether that is the case or not. I left it in for now; it doesn't seem broken down.