Fix bug where output was always required to be constructed in the Config.
Some clap options/flags do not require an output file to be given as argument.
The way Config is constructed, before running any processor, and including
the output path in the config, an output path was always required.
This commit makes Config.output an Option, so that one can be provided
if required, but also None can be given.