Reintegrate cli (Clap) related code within sic.
Separation of the cli (Clap) related code to stew_cli reduced code
duplication related to the cli code across sic and stew.
However, any change or addition to this module becomes very tedious,
especially since this module is currently constantly undergoing change
in order to simplify and optimise usage of sic (and stew).

For this reason, the module will be reintegrated within sic for the
foreseeable future, at least until the cli interface stabilises.
Then a decision will be made on what can be logically split to the cli
module, and what should be kept separated (even if this means duplication).