Add support for 'resize sampling_filter' with 'setopt' (#102)
Grammar for setopt (and setflag):
setflag [for-command] [flag-to-be-set];
setopt [for-comand] [option-to-be-set] [option-value];

* Adapt parsing module to support the new image transformation engine.
* Update parser grammar to allow for flags and options to be set.
* Update parser to support setopt resize sampling_filter.
* Reintegrate cli (Clap) related code within sic.

Separation of the cli (Clap) related code to stew_cli reduced code
duplication related to the cli code across sic and stew.
However, any change or addition to this module becomes very tedious,
especially since this module is currently constantly undergoing change
in order to simplify and optimise usage of sic (and stew).

For this reason, the module will be reintegrated within sic for the
foreseeable future, at least until the cli interface stabilizes.
Then a decision will be made on what can be logically split to the cli
module, and what should be kept separated (even if this means duplication).