Let Config contain the parsed image operations program.
Since we we'll be parsing image operation cli commands differently than the image operation commands from --apply-operations, it is useful to save the parsed Program to the config.
Then we can output a Program for the cli image operations as well, ready to be applied.

If we didn't saved the parsed program, we would need a new (second for image programs) config data field or output a string equivalent to what a user would have provided for an equivalent image operations program.

This change involves moving the Program within the Config.
To deal with the ownership of the program, we let the image engine take the Statements by reference instead of value.
The values requires for the image operations all implement Copy.