Add ability to parse value arguments separately from parser Rules.
Refines the image operation engine to allow better integration between
the by-rule and by-argument-values parsers.
That is, let the rule parser use the value parser for image operation arguments.

There will certainly be parts which can refined and integrated further,
however, as a first step of integration between the separate systems for
users to specify image operations, I'm quite happy.

The current solutions however, while the Rule parser does re-use some  parts
of the arg value parser, do exist 'next to each other'. Perhaps we could
improve on it by making all operations traits, and let them depend on
traits for parsing of values directly or by Rule; we'll have to think
about making it modular, so if we ever want to replace Pest with another parsing
system, we could without hopefully breaking too much.
Specifically, I would love the tests not to rely on these implementation
systems. If we replace Pest, we'll have to rewrite quite a few test cases.
If instead an opaque trait was used, perhaps the user wouldn't have to care
it is implemented using Pest in the back-end.
Something to spend some thought about in the future :).