Add all image operations as cli arguments. (#144)
* Adds image operations and operation modifiers as cli arguments.

'sic' now allows one to specify image operation through cli arguments. Previously, all operations had to be specified through the '--apply-operations' script (AOS). Letting one specify them as cli arguments feels more natural, and has as added bonus that --help does list all available operations.
The new method is currently called Image Operations as CLI Arguments (IOCA).


Arguments are handled by Clap, which returns them as iterable values by their argument name. For each value we can also request its index. Recall (or think about it for a minute 😄) that the order in which operations are applied does matter in some cases. The availability of indices is very helpful as it allows us to determine the order of the provided image operation cli arguments.  However, since indices related to some operation can be provided in any order, and with many other operations in between, we do have to do some processing to ensure we agree what value belongs to a certain operation, and what we should allow and reject.
For example, '--blur 1 --crop 0 0 3 3 --inverse' is correct, but according to clap, '--blur 1 --crop 0 0 --inverse --crop 3 3' is equivalent, as 4 values which are required for crop are provided, and since we want to be able to repeat operations, we have to allow multiple arguments. 
Thus, we also use indices to check that '--crop 0 0 3 3' is valid (i.e. has incremental argument values), but that '--crop 0 0 --crop 3 3' is not and should be rejected.
Additionally, since we receive an iterator of values for a specific cli argument (by its name), practically seen we receive for crop in the example 'sic -i some -o some2 --crop 0 0 3 3 --crop 0 0 2 2', the string values [0, 0, 3, 3, 0, 0, 2, 2]. and indices [6, 7, 8, 9, 11, 12, 13, 14]. We know that crop always takes 4 argument values, and use it to partition the iterator we received. Using the partitions we have the groups required arguments for each crop 'invocation'.
Then we proceed and parse the actual values and if that succeed build an image operation program.


* Config has been modified to contain the parsed image operations program (instead of the unparsed script).

Since we we'll be parsing image operation cli commands differently than the image operation commands from --apply-operations, it is useful to save the parsed Program to the config.
Then we can output a Program for the cli image operations as well, ready to be applied.

If we didn't saved the parsed program, we would need a new (second for the image program) config data field or output a string equivalent to what a user would have provided for an equivalent image operations program.

This change involves moving the Program within the Config.
To deal with the ownership of the program, we also let the image engine take the Statements by reference instead of value. The values requires for the image operations all implement Copy (they are primitive data types like i32, f32 and u32).

* Common test includes were moved to its own module (to avoid some repetition of setting up test cases). Thanks rust book: https://doc.rust-lang.org/book/ch11-03-test-organization.html

* Some effort has been made to integrate and modularize the existing parser used by AOS.
This parser operated on Pairs and Rules received from the Pest parser. For the IOCA method we don't use Pairs, instead, we parse the values of arguments directly. To re-use the value parsing which had to happen with the pair/rule parser as well, that part was made more modular. 

The image operation engine was refined to allow better integration between
the by-rule and by-argument-values parsers.
That is, the rule parser now uses the value parser for image operation arguments.

There will certainly be parts which can refined and integrated further,
however, as a first step of integration between the separate systems for
users to specify image operations, I'm quite happy.

The current solutions however, while the Rule parser does re-use some parts
of the arg value parser, do exist 'next to each other'. Perhaps we could
improve on it by making all operations traits, and let them depend on
traits for parsing of values directly or by Rule; we'll have to think
about making it modular, so if we ever want to replace Pest with another parsing
system, we could without hopefully breaking too much.
Specifically, I would love the tests not to rely on these implementation
systems. If we replace Pest, we'll have to rewrite quite a few test cases.
If instead an opaque trait was used, perhaps the user wouldn't have to care
it is implemented using Pest in the back-end.
Something to spend some thought about in the future :).

* Update the readme for image operations through cli arguments.