Add the ability to extract a custom frame from an animated gif instead of always the first.
**Overview:**
If an image is an animated image, that is, the image consist of multiple frames which can be played in some order, we currently always drop all frames except for the first. The image buffer will thus contain a still image.

Eventually we would like to provide support for animated images.
This PR is the first step to support animated images. It deals with selecting a custom frame which will be extracted to our (still) image buffer.

A user can select a frame by using the --select-frame <FRAME> argument.
<FRAME> allowed values:
* one-indexed unsigned integer  (1..usize::MAX): to select a frame by the index
* first: to select the first frame
* last: to select the last frame

Importing (loading) images was rewritten to accommodate for this change. The implementation has been simplified, and specifically now uses the `Read` trait instead of using indirections like `ExportMethod`. The images are now also always first read into a buffer and then imported as image from the buffer, instead of:
* if we have a path: open from the path (by using the extension to decide on the format) OR
* if we have no path: import from memory (by using headers to guess the format)
with the limitation that if an image had no extension (or an incorrect one), we would reject the input image.
Of course, always guessing the format by its header may also result in similar limitations.