Replace string errors with thiserror crate for the image engine.
This is the first sic component where we'll replace the string based
error handling. Strings are not great for error handling since they
prevent many diagnostic decorations.
If we were to use the std library error handling we would be implementing
std::error::Error and std::fmt::Display (and sometimes std::convert::From).
This can quickly become a verbose task, but it is also straightforward,
easy to maintain and works well for compatibility.

To reduce verbosity I've selected the thiserror crate which is compatible
with the std library method. This crate provides attribute macros which
help with implementing Display, Error and From amongst others.

The bin part of the project will use the anyhow crate by the same author.
Anyhow helps by providing additional (stacked) contexts and is used
instead of Box<dyn Error>. Anyhow has not been used yet (except for
an added dependency).

More about error handling can be found here:
https://blog.burntsushi.net/rust-error-handling/

The article is a few years old, yet still relevant for error handling using
the std library.