Use the thiserror library for error handling in the image i/o module.
Previously, various error handling strategies were used:
* String
* enum based implementing std::error:Error
* Box<dyn Error>

The bin part of the project has been made compatible, so we can
run integration tests and run the program itself, however,
it will require a revision to make it behave properly with
the anyhow library. The current version contains ad hoc
anyhow! errors with TODO messages for the places which need to
be updated.