Implement diff operation which shows the pixel differences between two images.
If a pixel at `(x, y)` in the image `this` (`P`) compared to the pixel at `(x, y)` in the image `that` (`Q`):
* is the same: the output image will colour that pixel white.
* differs: the output image will colour that pixel red.

The output image (`R`) will have width `w=max(width(this), width(that))` and height
`h=max(height(this), height(that))`.

In case that two images when overlapped differ inversely in both width and height, so
`(P_width > Q_width ∧ P_height < Q_height) ⊕ (P_width < Q_width ∧ P_height > Q_height)` then
there will be pixels in `R`, for which for some pixels `p_{i, j} ∈ R | p_{i, j} ∉ P ∨ p_{i, j} ∉ Q`.
That is, the part of output image which isn't part of either of the two original input images.
These pixels will be 'coloured' black but with an alpha value of 0, so they will be transparent
as to show they were not part of the input images.