Set the default sampling filter for resize to Lanczos3.
Lanczos3 produces subjectively an improved image quality (mostly sharper) compared to the current default,
which uses a Gaussian filter.
According to a single test case located in the image docs [1], Lanczos3 would have a similar computing
cost compared a Gaussian filter. This cost is relatively higher than nearest neighbor, triangle or
a Catmull-Rom filter, but produces subjectively the best quality images. For now we will opt to use
the "highest available quality" and we can use the resize_sampling_filter option to use another filter
if so desired (for example if the computing cost would be too high).
Admittedly, this switch is based on some subjectivity and assumptions. We should aim to test, benchmark
and properly study these sampling filters in the future.

[1] https://docs.rs/image/0.23.4/image/imageops/enum.FilterType.html

closes #339