pipeline::run depended on both argmatches and config while it should only use locally known data
This commit removes ArgMatches as an argument of pipeline::run. The aim is that it will enable us in the future
to run the pipeline in parallel with different inputs/outputs (which now becomes possible by setting the config
in- and output to different values, e.g. by allowing the input and output cli arguments to take multiple values).

This is an intermediate step, so we don't need to refactor and optimize the whole pipeline just yet.