Add implementation for diff using image script
Contains a bit of a hack for the diff trough the cli option, as a shell already removes the quotation marks, but our
script version does contain them. I wasn't able so far to get Pest to not include them as parse tokens (so this will be
future work for now).