Add the option to fallback to image crate format resolution
Useful when sic doesn't support an extension yet, but we have already upgraded to a newer version of the image crate which does. We still need out own methods for image output formats which accept different configurations (e.g. we allow to set the jpeg quality and pnm type explicitly, while the image crate uses sane defaults for these convenience functions). It also allows us to define our own identifiers and extensions if we wish :).