Refactor cli image ops parser
The previous version used Clap's ArgMatches (the output of Clap's parsing and validation). This was suboptimal, because it's output isn't meant for re-parsing of arguments. The new parser still relies on Clap for some validation (since that happens anyways; but is good to know about! :)), but uses the cli arguments given to the program directly instead of using ArgMatches for the processing of given arguments. The new parser should make it easier to add new image operations as cli arguments.