Support filtering glob paths on known extensions
The current implementation doesn't filter files (it does filter folders though!). What this means is that if a (glob) path contains an unsupported file path, eventually an unsupported file format error will be thrown. By pre-filtering on known image file extensions, we attempt to prevent having a user specify very complex and specific globs.

This implementation re-uses "DetermineEncodingFormat.by_extension(ext)" to check whether an extension is known. A disadvantage of this approach is that the above trait methods does slightly more than checking just whether the extension is supported (it also constructs an encoding format). Additionally, now this trait method will be run twice. Once here, and once when determining the format to save it to disk. Perhaps it can be cached in the future?

The implementation handles both the extensions supported directly by sic, as well as the extensions as defined by the image crate fallback decider.