Fix adding support for AVIF through the image crate
This is a bit of a hack, since we depend on the ImageOutputFormat which did not yet have an Avif entry for image crate v0.23.10. This has been contributed, but it is not released yet. Thus we have to call the encoder separately, which would require quite a bit of refactoring for a simple change (or we could wait I suppose), which would then revert the refactoring. For now however, I decided to add support, but use a environment variable to signal that the ImageOutputFormat is actually not Farb when the id "avif" is found, but Avif instead. Obviously this should be fixed when image crate `>v0.23.10` will be released.