Add the skeleton of the future "duo" parser
This parser is intended tp replace both sic_parser and sic_cli_ops, unifying the parsing experience. This is intended to make it easier to add new image operations, removing duplication through the code base.
As a disadvantage, we'll need to deal with two different input methods, which was the original reason two parsers exist: scripts in text form and cli ops through the cli interface.

This commit has removed the cli_ops_parser, but the sic_parser is still in. Eventually, that one should also be removed. We will simultaneously develop and test both user facing interfaces, and should setup compatibility testing against the latest version which uses the old parsers.