Increase the default time between publishing to 15 seconds
We wait between publishing so the package registry index has time to update, before our next fetch from it. The wait time was previously 5 seconds, but even when waiting 10 seconds the registry wasn't updated fast enough. We'll now attempt to default to 15 seconds. If that's not enough perhaps we'll have to either increase it further, or implement a retry strategy.