Keep dep licenses outside binary for reduced file size
Release size is 7.1MB as guessed by cargo-bloat. Previously was 10.8MB.
While this is significant, I'm not sure whether saving a few MB is worth it currently. If we want to further optimise file size, without it costing performance and while retaining symbols, we would probably have to look into reducing the amount of generics.