Add sic_imageproc crate
This crate will contain functions and filters ported from `imageproc` but using `image 0.24`.

Imageproc, which we used for slightly more advanced image operations has a hard dependency on the image primitives in the image crate.

In the best case scenario, imageproc would be in sync with image. This unfortunately isn't the case, and I currently have neither enough knowledge  nor enough time to learn the internals of the imageproc crate. In addition, it's difficult to port imageproc incrementally, since imageproc is tightly coupled to the types defined by image.
 As a result, I'm probably not the best person to take on porting imageproc to image 0.24.

Sic doesn't use most of imageproc, which allows me to just port over the parts I do use, and do so incrementally; this will be the purpose of the sic_imageproc crate.