Fix issue where image could not be loaded from stdin
Previously we would explicitely require the --input or --glob-input argument to be present, for the input mode to be selected (batch or single file input and output).

Now we instead check if --glob-input is present, and if it is select batch mode, otherwise we always select 'simple' mode (i.e. single input and output). NB: this also means that stdin should always use single input and output.